arrange test case
